changed database structures , measurements now dont have any plant , on plant deletions cascade works correctly and application is a quick boy again : ) ,PULL DATA ASWELL

+++ src/main/java/at/qe/skeleton/ui/controllers/PlantController.java
@@ -40,7 +40,7 @@ public class PlantController implements Serializable{
     /**
      * Returns how many plants are registered in the system.
      */
-    public Integer getPlantsAmount() {
+    public Long getPlantsAmount() {
         return plantService.getPlantsAmount();
     }
